Energy Test of Mutual Independence

The test statistic is the sum of d-1 bias-corrected squared dcor statistics where the number of variables is d. Implementation is by permuation test.

mutualIndep.test(x, R)

Arguments

x

data matrix or data frame

R

number of permutation replicates

Details

A population coefficient for mutual independence of d random variables, \(d \geq 2\), is $$ \sum_{k=1}^{d-1} \mathcal R^2(X_k, [X_{k+1},\dots,X_d]). $$ which is non-negative and equals zero iff mutual independence holds. For example, if d=4 the population coefficient is $$ \mathcal R^2(X_1, [X_2,X_3,X_4]) + \mathcal R^2(X_2, [X_3,X_4]) + \mathcal R^2(X_3, X_4), $$ A permutation test is implemented based on the corresponding sample coefficient. To test mutual independence of $$X_1,\dots,X_d$$ the test statistic is the sum of the d-1 statistics (bias-corrected \(dcor^2\) statistics): $$\sum_{k=1}^{d-1} \mathcal R_n^*(X_k, [X_{k+1},\dots,X_d])$$.

Value

mutualIndep.test returns an object of class power.htest.

Note

See Szekely and Rizzo (2014) for details on unbiased \(dCov^2\) and bias-corrected \(dCor^2\) (bcdcor) statistics.

See also

References

Szekely, G.J., Rizzo, M.L., and Bakirov, N.K. (2007), Measuring and Testing Dependence by Correlation of Distances, Annals of Statistics, Vol. 35 No. 6, pp. 2769-2794.
doi:10.1214/009053607000000505

Szekely, G.J. and Rizzo, M.L. (2014), Partial Distance Correlation with Methods for Dissimilarities. Annals of Statistics, Vol. 42 No. 6, 2382-2412.

Author

Maria L. Rizzo mrizzo@bgsu.edu and Gabor J. Szekely

Examples

x <- matrix(rnorm(100), nrow=20, ncol=5)
mutualIndep.test(x, 199)
#> 
#>      Energy Test of Mutual Independence 
#> 
#>       statistic = -0.09018846
#>         p.value = 0.66
#>            call = mutualIndep.test(x = x, R = 199)
#>       data.name = x  dim  20,5
#>        estimate = -0.060, -0.025, -0.024, 0.019
#> 
#> NOTE: statistic=sum(bcdcor); permutation test
#>