Inner product in the Hilbert space of U-centered distance matrices

Stand-alone function to compute the inner product in the Hilbert space of U-centered distance matrices, as in the definition of partial distance covariance.

U_product(U, V)

Arguments

U

U-centered distance matrix

V

U-centered distance matrix

Details

Note that pdcor, etc. functions include the centering and projection operations, so that these stand alone versions are not needed except in case one wants to check the internal computations.

Exported from U_product.cpp.

Value

U_product returns the inner product, a scalar.

References

Szekely, G.J. and Rizzo, M.L. (2014), Partial Distance Correlation with Methods for Dissimilarities, Annals of Statistics, Vol. 42, No. 6, pp. 2382-2412.
https://projecteuclid.org/euclid.aos/1413810731

Author

Maria L. Rizzo mrizzo@bgsu.edu and Gabor J. Szekely

Examples

 x <- iris[1:10, 1:4]
 y <- iris[11:20, 1:4]
 M1 <- as.matrix(dist(x))
 M2 <- as.matrix(dist(y))
 U <- U_center(M1)
 V <- U_center(M2)

 U_product(U, V)
#> [1] 0.01008021
 dcovU_stats(M1, M2)
#>      dCovU     bcdcor     dVarXU     dVarYU 
#> 0.01008021 0.15335234 0.03774546 0.11447062