Distance Correlation t-Test

Distance correlation t-test of multivariate independence for high dimension.

dcorT.test(x, y)
dcorT(x, y)

Arguments

x

data or distances of first sample

y

data or distances of second sample

Details

dcorT.test performs a nonparametric t-test of multivariate independence in high dimension (dimension is close to or larger than sample size). As dimension goes to infinity, the asymptotic distribution of the test statistic is approximately Student t with \(n(n-3)/2-1\) degrees of freedom and for \(n \geq 10\) the statistic is approximately distributed as standard normal.

The sample sizes (number of rows) of the two samples must agree, and samples must not contain missing values.

The t statistic (dcorT) is a transformation of a bias corrected version of distance correlation (see SR 2013 for details).

Large values (upper tail) of the dcorT statistic are significant.

Note

dcor.t and dcor.ttest are deprecated.

Value

dcorT returns the dcor t statistic, and dcorT.test returns a list with class htest containing

method

description of test

statistic

observed value of the test statistic

parameter

degrees of freedom

estimate

(bias corrected) squared dCor(x,y)

p.value

p-value of the t-test

data.name

description of data

See also

References

Szekely, G.J. and Rizzo, M.L. (2013). The distance correlation t-test of independence in high dimension. Journal of Multivariate Analysis, Volume 117, pp. 193-213.
doi: 10.1016/j.jmva.2013.02.012

Szekely, G.J., Rizzo, M.L., and Bakirov, N.K. (2007), Measuring and Testing Dependence by Correlation of Distances, Annals of Statistics, Vol. 35 No. 6, pp. 2769-2794.
doi: 10.1214/009053607000000505

Szekely, G.J. and Rizzo, M.L. (2009), Brownian Distance Covariance, Annals of Applied Statistics, Vol. 3, No. 4, 1236-1265.
doi: 10.1214/09-AOAS312

Author

Maria L. Rizzo mrizzo@bgsu.edu and Gabor J. Szekely

Examples

 x <- matrix(rnorm(100), 10, 10)
 y <- matrix(runif(100), 10, 10)
 dcorT(x, y)
#> [1] 0.0601793
 dcorT.test(x, y)
#> 
#> 	dcor t-test of independence for high dimension
#> 
#> data:  x and y
#> T = 0.060179, df = 34, p-value = 0.4762
#> sample estimates:
#> Bias corrected dcor 
#>          0.01032012 
#>