Fast dCor and dCov for bivariate data only

For bivariate data only, these are fast O(n log n) implementations of distance correlation and distance covariance statistics. The U-statistic for dcov^2 is unbiased; the V-statistic is the original definition in SRB 2007. These algorithms do not store the distance matrices, so they are suitable for large samples.

dcor2d(x, y, type = c("V", "U"))
dcov2d(x, y, type = c("V", "U"), all.stats = FALSE)

Arguments

x
numeric vector

y
numeric vector

type
"V" or "U", for V- or U-statistics

all.stats
logical

Details

The unbiased (squared) dcov is documented in dcovU, for multivariate data in arbitrary, not necessarily equal dimensions. dcov2d and dcor2d provide a faster O(n log n) algorithm for bivariate (x, y) only (X and Y are real-valued random vectors). The O(n log n) algorithm was proposed by Huo and Szekely (2016). The algorithm is faster above a certain sample size n. It does not store the distance matrix so the sample size can be very large.

Value

By default, dcor2d returns the V-statistic dCor_n^2(x, y), and if type="U", it returns a bias-corrected estimator of squared dcor.

By default, dcov2 returns the V-statistic V_n^2 = dCov_n^2(x, y), and if type="U", it returns the U-statistic, unbiased for V^2(X, Y). The argument all.stats=TRUE is used internally when the function is called from dcor2.

For dcov2d and dcor2d, direct computation using the C++ function dcovU_stats may be somewhat faster on small samples, depending on the platform.

dcor2d and dcov2d do not store the distance matrices so these functions are helpful when sample size is large, the data is bivariate, and we simply require the statistics. There is not an efficient way to do the nonparametric test by permutations without storing distances. For a test of independence on moderate size samples, use dcov.test or dcor.test.

Note

Unbiased distance covariance (SR2014) is equivalent to the U-statistic estimator of \(\mathrm{dCov^2}\). Since dcovU is an unbiased statistic, it can be negative and its square root would be complex so the square root of the U-statistic is not applied. For the original distance covariance test of independence (SRB2007, SR2009), the sample distance covariance was defined by the square root of the V-statistic \(\mathrm{n\, dCov_n^2} = \mathcal{V}_n^2\). Similarly, bcdcor is bias-corrected, so we do not take the square root as with \(dCor_n^2\).

References

Huo, X. and Szekely, G.J. (2016). Fast computing for distance covariance. Technometrics, 58(4), 435-447.

Szekely, G.J. and Rizzo, M.L. (2014), Partial Distance Correlation with Methods for Dissimilarities. Annals of Statistics, Vol. 42 No. 6, 2382-2412.

Szekely, G.J., Rizzo, M.L., and Bakirov, N.K. (2007), Measuring and Testing Dependence by Correlation of Distances, Annals of Statistics, Vol. 35 No. 6, pp. 2769-2794.
http://dx.doi.org/10.1214/009053607000000505

Examples

  

    ## these are equivalent, but 2d is faster for n > 50
    n <- 100
    x <- rnorm(100)
    y <- rnorm(100)
    all.equal(dcov(x, y)^2, dcov2d(x, y), check.attributes = FALSE)

#> [1] TRUE

    all.equal(bcdcor(x, y), dcor2d(x, y, "U"), check.attributes = FALSE)

#> [1] TRUE