Energy Statistic Test of Independence

Deprecated: use indep.test with method = mvI. Computes a multivariate nonparametric E-statistic and test of independence.

indep.e(x, y)
indep.etest(x, y, R)

Arguments

x
matrix: first sample, observations in rows

y
matrix: second sample, observations in rows

R
number of replicates

Details

Computes the coefficient \(\mathcal I\) and performs a nonparametric \(\mathcal E\)-test of independence. The test decision is obtained via bootstrap, with R replicates. The sample sizes (number of rows) of the two samples must agree, and samples must not contain missing values. The statistic \(\mathcal E = n \mathcal I^2\) is a ratio of V-statistics based on interpoint distances \(\|x_{i}-y_{j}\|\). See the reference below for details.

Value

The sample coefficient \(\mathcal I\) is returned by indep.e. The function indep.etest returns a list with class htest containing

method

description of test

statistic

observed value of the coefficient \(\mathcal I\)

p.value

approximate p-value of the test

data.name

description of data

References

Bakirov, N.K., Rizzo, M.L., and Szekely, G.J. (2006), A Multivariate Nonparametric Test of Independence, Journal of Multivariate Analysis 93/1, 58-80,
http://dx.doi.org/10.1016/j.jmva.2005.10.005

Examples

 
# NOT RUN {
 ## independent univariate data
 x <- sin(runif(30, 0, 2*pi) * 2)
 y <- sin(runif(30, 0, 2*pi) * 4)
 indep.etest(x, y, R = 99)

 ## dependent multivariate data
 Sigma <- matrix(c(1, .1, 0, 0 , 1, 0, 0 ,.1, 1), 3, 3)
 x <- mvrnorm(30, c(0, 0, 0), diag(3))
 y <- mvrnorm(30, c(0, 0, 0), Sigma) * x
 indep.etest(x, y, R = 99)

# }