Energy Statistic Test of Independence

Computes the multivariate nonparametric E-statistic and test of independence based on independence coefficient \(\mathcal I_n\).

mvI.test(x, y, R)
    mvI(x, y)

Arguments

x
matrix: first sample, observations in rows

y
matrix: second sample, observations in rows

R
number of replicates

Details

Computes the coefficient \(\mathcal I\) and performs a nonparametric \(\mathcal E\)-test of independence. The test decision is obtained via bootstrap, with R replicates. The sample sizes (number of rows) of the two samples must agree, and samples must not contain missing values. The statistic \(\mathcal E = n \mathcal I^2\) is a ratio of V-statistics based on interpoint distances \(\|x_{i}-y_{j}\|\). See the reference below for details.

Value

mvI returns the statistic. mvI.test returns a list with class htest containing

method

description of test

statistic

observed value of the test statistic \(n\mathcal I_n^2\)

estimate

\(\mathcal I_n\)

replicates

replicates of the test statistic

p.value

approximate p-value of the test

data.name

description of data

References

Bakirov, N.K., Rizzo, M.L., and Szekely, G.J. (2006), A Multivariate Nonparametric Test of Independence, Journal of Multivariate Analysis 93/1, 58-80,
http://dx.doi.org/10.1016/j.jmva.2005.10.005

Note

Historically this is the first energy test of independence. The distance covariance test dcov.test, distance correlation dcor, and related methods are more recent (2007,2009). The distance covariance test is faster and has different properties than mvI.test. Both methods are based on a population independence coefficient that characterizes independence and both tests are statistically consistent.

Author

Maria L. Rizzo mrizzo @ bgsu.edu and Gabor J. Szekely

See also