Partial distance correlation and covariance

Partial distance correlation pdcor, pdcov, and tests.

pdcov.test(x, y, z, R)
  pdcor.test(x, y, z, R)
  pdcor(x, y, z)
  pdcov(x, y, z)

Arguments

x

data matrix or dist object of first sample

y

data matrix or dist object of second sample

z

data matrix or dist object of third sample

R

replicates for permutation test

Details

pdcor(x, y, z) and pdcov(x, y, z) compute the partial distance correlation and partial distance covariance, respectively, of x and y removing z.

A test for zero partial distance correlation (or zero partial distance covariance) is implemented in pdcor.test, and pdcov.test.

If the argument is a matrix, it is treated as a data matrix and distances are computed (observations in rows). If the arguments are distances or dissimilarities, they must be distance (dist) objects. For symmetric, zero-diagonal dissimilarity matrices, use as.dist to convert to a dist object.

Value

Each test returns an object of class htest.

Author

Maria L. Rizzo mrizzo @ bgsu.edu and Gabor J. Szekely

References

Szekely, G.J. and Rizzo, M.L. (2014), Partial Distance Correlation with Methods for Dissimilarities. Annals of Statistics, Vol. 42 No. 6, 2382-2412.

Examples

  n = 30
  R <- 199

  ## mutually independent standard normal vectors
  x <- rnorm(n)
  y <- rnorm(n)
  z <- rnorm(n)

  pdcor(x, y, z)
#>     pdcor 
#> 0.1550543 
  pdcov(x, y, z)
#> [1] 0.07958858
  set.seed(1)
  pdcov.test(x, y, z, R=R)
#> 
#> 	pdcov test
#> 
#> data:  replicates 199
#> n V^* = 2.3877, p-value = 0.02
#> sample estimates:
#>     pdcor 
#> 0.1550543 
#> 
  set.seed(1)
  pdcor.test(x, y, z, R=R)
#> 
#> 	pdcor test
#> 
#> data:  replicates 199
#> pdcor = 0.15505, p-value = 0.02
#> sample estimates:
#>     pdcor 
#> 0.1550543 
#> 

# \donttest{
  if (require(MASS)) {
    p = 4
    mu <- rep(0, p)
    Sigma <- diag(p)
  
    ## linear dependence
    y <- mvrnorm(n, mu, Sigma) + x
    print(pdcov.test(x, y, z, R=R))
  
    ## non-linear dependence
    y <- mvrnorm(n, mu, Sigma) * x
    print(pdcov.test(x, y, z, R=R))
    }
#> 
#> 	pdcov test
#> 
#> data:  replicates 199
#> n V^* = 12.934, p-value = 0.005
#> sample estimates:
#>     pdcor 
#> 0.6411337 
#> 
#> 
#> 	pdcov test
#> 
#> data:  replicates 199
#> n V^* = 0.59727, p-value = 0.125
#> sample estimates:
#>      pdcor 
#> 0.05074763 
#> 
  # }