Mean Distance Test for Poisson Distribution

Performs the mean distance goodness-of-fit test of Poisson distribution with unknown parameter.

poisson.mtest(x, R)
poisson.m(x)

Arguments

x
vector of nonnegative integers, the sample data

R
number of bootstrap replicates

Details

The mean distance test of Poissonity was proposed and implemented by Szekely and Rizzo (2004). The test is based on the result that the sequence of expected values E|X-j|, j=0,1,2,... characterizes the distribution of the random variable X. As an application of this characterization one can get an estimator \(\hat F(j)\) of the CDF. The test statistic (see poisson.m) is a Cramer-von Mises type of distance, with M-estimates replacing the usual EDF estimates of the CDF: $$M_n = n\sum_{j=0}^\infty (\hat F(j) - F(j\;; \hat \lambda))^2 f(j\;; \hat \lambda).$$ The test is implemented by parametric bootstrap with R replicates.

Value

The function poisson.m returns the test statistic. The function poisson.mtest returns a list with class htest containing

method

Description of test

statistic

observed value of the test statistic

p.value

approximate p-value of the test

data.name

description of data

estimate

sample mean

References

Szekely, G. J. and Rizzo, M. L. (2004) Mean Distance Test of Poisson Distribution, Statistics and Probability Letters, 67/3, 241-247. http://dx.doi.org/10.1016/j.spl.2004.01.005.

Examples

 x <- rpois(20, 1)
 poisson.m(x)

#> [1] 0.01206805

 poisson.mtest(x, R = 199)

#> 
#> 	Mean distance test of Poisson distribution
#> 
#> data:  sample size 20, replicates 199
#> test statistic = 0.012068, p-value = 0.7487
#> sample estimates:
#> mean 
#> 0.95 
#>